refactor: Move auth route tests to auth folder
