+++ test/api/v3/integration/user/auth/PUT-user_update_email.test.js
@@ -1,8 +1,8 @@
 import {
   generateUser,
   translate as t,
-} from '../../../../helpers/api-v3-integration.helper';
-import { model as User } from '../../../../../website/src/models/user';
+} from '../../../../../helpers/api-v3-integration.helper';
+import { model as User } from '../../../../../../website/src/models/user';
 
 describe('PUT /user/auth/update-email', () => {
   let user;
